Bug Fixes
Falta corregir el bug de las fechas en los campos de empleados y clientes

+++ src/main/java/Funciones/Entidades/CRUDVideojuegos.java
@@ -32,7 +32,7 @@ public class CRUDVideojuegos {
                 int id = output.getInt("id_videojuego");
                 String nombre = output.getString("Nombre");
                 String categoria = output.getString("Categoria");
-                int precio = output.getInt("Precio");
+                double precio = output.getDouble("Precio");
 
                 data.add(new Videojuego(id, nombre, categoria, precio));
             }
